refactor: rename package to scdocbuilder
