+++ src/scdocbuilder/__main__.py
@@ -1,4 +1,4 @@
-"""Package entrypoint for ``python -m faa_sc_replacer``."""
+"""Package entrypoint for ``python -m scdocbuilder``."""
 
 from .cli import main
