+++ src/scdocbuilder/config.py
