+++ src/scdocbuilder/validation.py
