+++ tests/test_io.py
@@ -5,7 +5,7 @@ from pathlib import Path
 if not typing.TYPE_CHECKING:
     pytest.importorskip("docx")
 
-from faa_sc_replacer.io import validate_input_files
+from scdocbuilder.io import validate_input_files
 
 
 def test_validate_input_files(tmp_path: Path) -> None:
